Send minimized form HTML to model
Previously, we were sending a JSON representation of the form, which
made too many assumptions. Sending the HTML allows the AI to handle
various cases were the form HTML might not adhere to best practices.

We still need to update the field population logic to handle <select>
elements, but that will be implemented in an upcoming commit.